backend-ai/requirements.txt
websockets==15.0.1
google-generativeai==0.8.5
python-dotenv==1.2.1

